CPM-757 : refacto AutoNumberLimitPerStructure
