+++ components/identifier-generator/back/tests/Specification/Infrastructure/Validation/IdentifierGeneratorCreationLimitValidatorSpec.php
@@ -31,15 +31,6 @@ class IdentifierGeneratorCreationLimitValidatorSpec extends ObjectBehavior
         $this->shouldThrow(\InvalidArgumentException::class)->during('validate', ['code', new NotBlank()]);
     }
 
-    public function it_could_throw_an_error_when_its_not_the_right_command(ExecutionContext $context): void
-    {
-        $context->getRoot()
-            ->willReturn(new \stdClass());
-        $this->shouldThrow(\InvalidArgumentException::class)->during('validate', ['code', new IdentifierGeneratorCreationLimit(['limit' => 2])]);
-
-        $this->shouldThrow(\InvalidArgumentException::class)->during('validate', ['code', new IdentifierGeneratorCreationLimit()]);
-    }
-
     public function it_should_build_violation_when_an_identifier_generator_already_exist(
         ExecutionContext $context,
         IdentifierGeneratorRepository $repository
