+++ components/identifier-generator/back/tests/Specification/Infrastructure/Validation/TargetAttributeShouldExistValidatorSpec.php
@@ -33,13 +33,6 @@ class TargetAttributeShouldExistValidatorSpec extends ObjectBehavior
         $this->shouldThrow(\InvalidArgumentException::class)->during('validate', ['code', new NotBlank()]);
     }
 
-    public function it_could_throw_an_error_when_its_not_the_right_command(ExecutionContext $context): void
-    {
-        $context->getRoot()
-            ->willReturn(new \stdClass());
-        $this->shouldThrow(\InvalidArgumentException::class)->during('validate', ['code', new TargetAttributeShouldExist()]);
-    }
-
     public function it_should_build_violation_when_target_attribute_does_not_exist(ExecutionContext $context): void
     {
         $context->getRoot()
